Use OCaml 4.02.3 in Docker image
It seems that the OCaml compiler version of the default ocaml/opam
Docker image has recently changed from 4.02.3 to 4.03.0. (See
https://github.com/ocaml/opam-dockerfiles/commit/a323bfc99bce6718b713df27942292983d4eab2e)
At the moment we do not support 4.03.0, only 4.02.3.

This commit explicitly fixes the OCaml version to 4.02.3, and uses the
Debian-based image (the default image is also based on Debian):
ocaml/opam:debian-stable_ocaml-4.02.3
(Alternatively I could specify the Debian release:
ocaml/opam:debian-8_ocaml-4.02.3)

This is required to make the Travis CI build work, since we use this
Docker image for testing the repository with Travis.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>