Use older Lwt release that does not require Result
Use the older Lwt release for the packages that do not compile at the
moment. We do not yet support the newer result type required by the new
release of Lwt - our packages are not compatible with it at the moment.

Some of our packages that use Lwt work fine with the new 2.6 version of
Lwt, and I'm not sure why, so I only changed the ones that do not
compile.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>